refactor(compile): kill shuffle v1 and rename shuffleV2 -> shuffle
- Delete pkg/sql/colexec/shuffle/ (v1) directory and all references.
- Rename pkg/sql/colexec/shuffleV2/ -> shuffle/, ShuffleV2 -> Shuffle.
- Delete compileShuffleJoinV2 and compileShuffleGroupV2; unified
  compileShuffleJoin/compileShuffleGroup always use the shuffle path.
- Delete constructShuffleOperatorForJoin and constructShuffleArgForGroup
  (v1). Remove vm.Shuffle cases from constructDuplicateOperator and
  remoterun send/receive.
- Remove shuffleV2 bool parameter from constructShuffleJoinOP; always
  sets ShuffleIdx=-1 unconditionally.
- Remove compileJoin v2 inline guards.
- Drop shuffle package import.
