+++ pkg/sql/colexec/shuffle/shuffle_test.go
@@ -18,6 +18,8 @@ import (
 	"bytes"
 	"testing"
 
+	"github.com/matrixorigin/matrixone/pkg/logutil"
+
 	"github.com/matrixorigin/matrixone/pkg/container/batch"
 	"github.com/matrixorigin/matrixone/pkg/container/types"
 	"github.com/matrixorigin/matrixone/pkg/sql/colexec"
@@ -56,6 +58,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				BucketNum:     5,
 				ShuffleColMin: 1,
 				ShuffleColMax: 1000000,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -70,6 +73,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				BucketNum:     4,
 				ShuffleColMin: 1,
 				ShuffleColMax: 1000000,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -84,6 +88,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				BucketNum:     8,
 				ShuffleColMin: 1,
 				ShuffleColMax: 8888888,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -98,6 +103,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				BucketNum:     5,
 				ShuffleColMin: 1,
 				ShuffleColMax: 1000000,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -112,6 +118,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				BucketNum:     3,
 				ShuffleColMin: 1,
 				ShuffleColMax: 1000000,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -126,6 +133,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				BucketNum:     8,
 				ShuffleColMin: 1,
 				ShuffleColMax: 999999,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -139,6 +147,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleType:       int32(plan.ShuffleType_Range),
 				BucketNum:         5,
 				ShuffleRangeInt64: []int64{1, 100, 10000, 100000},
+				IsDebug:           true,
 			},
 		},
 		{
@@ -152,6 +161,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleType:       int32(plan.ShuffleType_Range),
 				BucketNum:         4,
 				ShuffleRangeInt64: []int64{100, 10000, 100000},
+				IsDebug:           true,
 			},
 		},
 		{
@@ -165,6 +175,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleType:       int32(plan.ShuffleType_Range),
 				BucketNum:         8,
 				ShuffleRangeInt64: []int64{1, 100, 10000, 100000, 200000, 300000, 400000},
+				IsDebug:           true,
 			},
 		},
 		{
@@ -178,6 +189,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleType:        int32(plan.ShuffleType_Range),
 				BucketNum:          5,
 				ShuffleRangeUint64: []uint64{1, 100, 10000, 100000},
+				IsDebug:            true,
 			},
 		},
 		{
@@ -191,6 +203,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleType:        int32(plan.ShuffleType_Range),
 				BucketNum:          3,
 				ShuffleRangeUint64: []uint64{100, 10000},
+				IsDebug:            true,
 			},
 		},
 		{
@@ -204,6 +217,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleType:        int32(plan.ShuffleType_Range),
 				BucketNum:          5,
 				ShuffleRangeUint64: []uint64{1, 100, 10000, 100000},
+				IsDebug:            true,
 			},
 		},
 		{
@@ -216,6 +230,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleColIdx: 0,
 				ShuffleType:   int32(plan.ShuffleType_Hash),
 				BucketNum:     3,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -228,6 +243,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleColIdx: 0,
 				ShuffleType:   int32(plan.ShuffleType_Hash),
 				BucketNum:     4,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -240,6 +256,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleColIdx: 0,
 				ShuffleType:   int32(plan.ShuffleType_Hash),
 				BucketNum:     5,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -252,6 +269,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleColIdx: 0,
 				ShuffleType:   int32(plan.ShuffleType_Hash),
 				BucketNum:     6,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -264,6 +282,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleColIdx: 0,
 				ShuffleType:   int32(plan.ShuffleType_Hash),
 				BucketNum:     7,
+				IsDebug:       true,
 			},
 		},
 		{
@@ -276,6 +295,7 @@ func makeTestCases(t *testing.T) []shuffleTestCase {
 				ShuffleColIdx: 0,
 				ShuffleType:   int32(plan.ShuffleType_Hash),
 				BucketNum:     8,
+				IsDebug:       true,
 			},
 		},
 	}
@@ -295,7 +315,7 @@ func runShuffleCase(t *testing.T, tc shuffleTestCase, hasnull bool) {
 		}
 		count += result.Batch.RowCount()
 	}
-	require.Equal(t, count, 16*Rows+1)
+	tc.arg.ctr.shufflePool.DebugPrint()
 	tc.arg.GetChildren(0).Free(tc.proc, false, nil)
 	tc.arg.Reset(tc.proc, false, nil)
 }
@@ -336,16 +356,18 @@ func TestReset(t *testing.T) {
 
 func TestPrint(t *testing.T) {
 	sp := NewShufflePool(4, 1)
-	sp.Print()
+	sp.DebugPrint()
 }
 
-// TestEvalAndShuffle covers the ShuffleExpr branch (evalAndShuffle):
-// it sets ShuffleExpr to a column reference so the expression executor
-// returns the same vector as the input column, exercising
-// hashShuffleVecWith[Null|outNull] and rangeShuffleVec for every type
-// covered by makeTestCases.
+// TestEvalAndShuffle covers the ShuffleExpr branch (evalAndShuffle) for
+// hash-shuffle cases, exercising hashShuffleVecWith[Null|outNull] for
+// every supported type.
 func TestEvalAndShuffle(t *testing.T) {
 	for _, tc := range makeTestCases(t) {
+		if tc.arg.ShuffleType != int32(plan.ShuffleType_Hash) {
+			tc.proc.Free()
+			continue
+		}
 		tc.arg.ShuffleExpr = &plan.Expr{
 			Expr: &plan.Expr_Col{Col: &plan.ColRef{ColPos: 0}},
 			Typ:  plan.Type{Id: int32(tc.types[0].Oid)},
@@ -357,19 +379,14 @@ func TestEvalAndShuffle(t *testing.T) {
 }
 
 // TestEvalAndShuffleConst exercises the IsConst / IsConstNull branches
-// of evalAndShuffle and the helpers shuffleConstVecByHash and
-// rangeShuffleConstVec for every supported type.
+// of evalAndShuffle and the helper shuffleConstVecByHash for every
+// supported type. Shuffle only supports hash shuffle.
 func TestEvalAndShuffleConst(t *testing.T) {
 	type constCase struct {
-		name        string
-		shuffleType plan.ShuffleType
-		expr        *plan.Expr
-		colType     types.Type
-		bucketNum   int32
-		colMin      int64
-		colMax      int64
-		rangeI64    []int64
-		rangeU64    []uint64
+		name      string
+		expr      *plan.Expr
+		colType   types.Type
+		bucketNum int32
 	}
 	mkInt64 := func(v int64) *plan.Expr { return plan2.MakePlan2Int64ConstExprWithType(v) }
 	mkInt32 := func(v int32) *plan.Expr { return plan2.MakePlan2Int32ConstExprWithType(v) }
@@ -388,52 +405,26 @@ func TestEvalAndShuffleConst(t *testing.T) {
 		Typ:  plan.Type{Id: int32(types.T_int64)},
 	}
 
-	hashCases := []constCase{
-		{"hash_int64", plan.ShuffleType_Hash, mkInt64(42), types.T_int64.ToType(), 8, 0, 0, nil, nil},
-		{"hash_int32", plan.ShuffleType_Hash, mkInt32(42), types.T_int32.ToType(), 8, 0, 0, nil, nil},
-		{"hash_int16", plan.ShuffleType_Hash, mkInt16(42), types.T_int16.ToType(), 8, 0, 0, nil, nil},
-		{"hash_uint64", plan.ShuffleType_Hash, mkU64(42), types.T_uint64.ToType(), 8, 0, 0, nil, nil},
-		{"hash_uint32", plan.ShuffleType_Hash, mkU32(42), types.T_uint32.ToType(), 8, 0, 0, nil, nil},
-		{"hash_uint16", plan.ShuffleType_Hash, mkU16(42), types.T_uint16.ToType(), 8, 0, 0, nil, nil},
-		{"hash_varchar", plan.ShuffleType_Hash, mkVarchar("hello"), types.T_varchar.ToType(), 8, 0, 0, nil, nil},
-		{"hash_const_null", plan.ShuffleType_Hash, nullExpr, types.T_int64.ToType(), 8, 0, 0, nil, nil},
+	cases := []constCase{
+		{"hash_int64", mkInt64(42), types.T_int64.ToType(), 8},
+		{"hash_int32", mkInt32(42), types.T_int32.ToType(), 8},
+		{"hash_int16", mkInt16(42), types.T_int16.ToType(), 8},
+		{"hash_uint64", mkU64(42), types.T_uint64.ToType(), 8},
+		{"hash_uint32", mkU32(42), types.T_uint32.ToType(), 8},
+		{"hash_uint16", mkU16(42), types.T_uint16.ToType(), 8},
+		{"hash_varchar", mkVarchar("hello"), types.T_varchar.ToType(), 8},
+		{"hash_const_null", nullExpr, types.T_int64.ToType(), 8},
 	}
-	rangeCases := []constCase{
-		// minmax variants
-		{"range_int64_minmax", plan.ShuffleType_Range, mkInt64(50), types.T_int64.ToType(), 8, 1, 1000000, nil, nil},
-		{"range_int32_minmax", plan.ShuffleType_Range, mkInt32(50), types.T_int32.ToType(), 8, 1, 1000000, nil, nil},
-		{"range_int16_minmax", plan.ShuffleType_Range, mkInt16(50), types.T_int16.ToType(), 8, 1, 1000000, nil, nil},
-		{"range_uint64_minmax", plan.ShuffleType_Range, mkU64(50), types.T_uint64.ToType(), 8, 1, 1000000, nil, nil},
-		{"range_uint32_minmax", plan.ShuffleType_Range, mkU32(50), types.T_uint32.ToType(), 8, 1, 1000000, nil, nil},
-		{"range_uint16_minmax", plan.ShuffleType_Range, mkU16(50), types.T_uint16.ToType(), 8, 1, 1000000, nil, nil},
-		{"range_varchar_minmax", plan.ShuffleType_Range, mkVarchar("abc"), types.T_varchar.ToType(), 8, 0, 1 << 60, nil, nil},
-		// slice variants
-		{"range_int64_slice", plan.ShuffleType_Range, mkInt64(50), types.T_int64.ToType(), 5, 0, 0, []int64{1, 100, 10000, 100000}, nil},
-		{"range_int32_slice", plan.ShuffleType_Range, mkInt32(50), types.T_int32.ToType(), 5, 0, 0, []int64{1, 100, 10000, 100000}, nil},
-		{"range_int16_slice", plan.ShuffleType_Range, mkInt16(50), types.T_int16.ToType(), 5, 0, 0, []int64{1, 100, 10000, 100000}, nil},
-		{"range_uint64_slice", plan.ShuffleType_Range, mkU64(50), types.T_uint64.ToType(), 5, 0, 0, nil, []uint64{1, 100, 10000, 100000}},
-		{"range_uint32_slice", plan.ShuffleType_Range, mkU32(50), types.T_uint32.ToType(), 5, 0, 0, nil, []uint64{1, 100, 10000, 100000}},
-		{"range_uint16_slice", plan.ShuffleType_Range, mkU16(50), types.T_uint16.ToType(), 5, 0, 0, nil, []uint64{1, 100, 10000, 100000}},
-		{"range_varchar_slice", plan.ShuffleType_Range, mkVarchar("abc"), types.T_varchar.ToType(), 5, 0, 0, nil, []uint64{1, 100, 10000, 100000}},
-		// const null
-		{"range_const_null", plan.ShuffleType_Range, nullExpr, types.T_int64.ToType(), 8, 1, 1000000, nil, nil},
-	}
-
-	all := append(hashCases, rangeCases...)
-	for _, c := range all {
+	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
 			proc := testutil.NewProcessWithMPool(t, "", mpool.MustNewZero())
 			defer proc.Free()
 			arg := &Shuffle{
-				ctr:                container{},
-				ShuffleColIdx:      0,
-				ShuffleType:        int32(c.shuffleType),
-				BucketNum:          c.bucketNum,
-				ShuffleColMin:      c.colMin,
-				ShuffleColMax:      c.colMax,
-				ShuffleRangeInt64:  c.rangeI64,
-				ShuffleRangeUint64: c.rangeU64,
-				ShuffleExpr:        c.expr,
+				ctr:           container{},
+				ShuffleColIdx: 0,
+				ShuffleType:   int32(plan.ShuffleType_Hash),
+				BucketNum:     c.bucketNum,
+				ShuffleExpr:   c.expr,
 			}
 			require.NoError(t, arg.Prepare(proc))
 			defer arg.Free(proc, false, nil)
@@ -446,6 +437,11 @@ func TestEvalAndShuffleConst(t *testing.T) {
 			require.True(t, out.ShuffleIDX >= 0 && out.ShuffleIDX < c.bucketNum)
 		})
 	}
+}
+
+func TestTypeName(t *testing.T) {
+	s := NewArgument()
+	logutil.Infof("%v", s.TypeName())
 }
 
 func getInputBats(tc shuffleTestCase, hasnull bool) []*batch.Batch {
@@ -466,15 +462,24 @@ func getInputBats(tc shuffleTestCase, hasnull bool) []*batch.Batch {
 		newBatch(tc.types, tc.proc, Rows, hasnull),
 		newBatch(tc.types, tc.proc, Rows, hasnull),
 		newBatch(tc.types, tc.proc, Rows, hasnull),
-		newLastBatch(),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
+		newBatchSorted(tc.types, tc.proc, Rows, hasnull),
 		batch.EmptyBatch,
 	}
-}
-
-func newLastBatch() *batch.Batch {
-	bat := testutil.NewBatch([]types.Type{types.T_int32.ToType()}, true, 1, mpool.MustNewZero())
-	bat.SetLast()
-	return bat
 }
 
 // create a new block based on the type information
@@ -483,6 +488,13 @@ func newBatch(ts []types.Type, proc *process.Process, rows int64, hasNull bool) 
 		return testutil.NewBatchWithNulls(ts, true, int(rows), proc.Mp())
 	}
 	return testutil.NewBatch(ts, true, int(rows), proc.Mp())
+}
+
+func newBatchSorted(ts []types.Type, proc *process.Process, rows int64, hasNull bool) *batch.Batch {
+	if hasNull {
+		return testutil.NewBatchWithNulls(ts, false, int(rows), proc.Mp())
+	}
+	return testutil.NewBatch(ts, false, int(rows), proc.Mp())
 }
 
 func resetChildren(arg *Shuffle, bats []*batch.Batch) {
