+++ pkg/sql/compile/debugTools.go
@@ -37,8 +37,7 @@ var debugInstructionNames = map[vm.OpType]string{
 	vm.ProductL2:               "product l2",
 	vm.Filter:                  "filter",
 	vm.Dispatch:                "dispatch",
-	vm.Shuffle:                 "shuffle",
-	vm.ShuffleV2:               "shuffleV2",
+	vm.Shuffle:               "shuffle",
 	vm.Connector:               "connect",
 	vm.Projection:              "projection",
 	vm.HashJoin:                "hash join",
