+++ pkg/vm/types.go
@@ -98,7 +98,6 @@ const (
 	LockOp
 
 	Shuffle
-	ShuffleV2
 
 	Sample
 	ProductL2
